backend/requirements.txt
# Backend Requirements - Django REST API
# SIM SMA BIMA

# Core Framework
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0

# API Documentation
drf-yasg==1.21.7
drf-spectacular==0.26.5

# Image handling
Pillow==10.1.0

# Date & Time
python-dateutil==2.8.2

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Data validation
django-phonenumber-field==7.2.0
phonenumbers==8.13.26

# File handling
django-cleanup==8.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Utilities
pytz==2023.3
